Fix broken Sugar.ConfigTest
Looks like the ordering of the results of get(:config_test) change depending on the Elixir version.  For
now, I've simply stuck the rhs in an if/then to use the new (as of 1.2.2) ordering.  I hope this is the
correct approach here.  I've yet to run into any actual issue with Sugar on Elixir 1.2.x, so I'm assuming
that this is just a testing issue.

It looks like this changed in elixir-lang/elixir/0c2f3a4 ("Document and remove Keyword dependency from
Dict"), in which the definition of Keyword.merge/2 swapped the left and right side (i.e. `keywords2 ++
:lists.filter(fun, keywords1)` became `:lists.filter(fun, keywords1) ++ keywords2`).