Fix mix.exs warnings and bump minimum Elixir version
We're now relying on a 1.2-specific feature (namely,
`Macro.camelize/1` and `Macro.underscore/1`), so this marks the formal
loss of support for Elixir 1.1 and older.  If support is requested for
some reason, we can probably put the use of `Macro` v. `Mix.Utils`
behind some kind of conditional, but otherwise, pre-1.2 users are
probably better off sticking to Sugar 0.4.x.